requirements.txt
flask
opencv-python
numpy
torch
torchvision
pillow
gunicorn